Create NormalizeAdditionSets method for future use.
This CL introduces the logic to "normalize" the addition sets provided
for the FPS Overrides enterprise policy.

All addition sets are ensured to be disjoint from one another by
FirstPartySetsHandler::ValidateEnterprisePolicy, however they may still
"transitively-overlap" by intersecting with the same existing First-
Party Set.

This "normalization" refers to unioning together any transitively
overlapping addition sets by using the Union-Find algorithm introduced
in https://crrev.com/c/3462488. The tie-breaker used for two addition
sets goes to whichever addition set was provided by the admin first.

After normalizing the addition sets, we can use the *new* normalized
addition sets to augment the existing list of sets.

Bug: 1244990, 1277128
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3591133
Commit-Queue: Kirubel Aklilu <kaklilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#995449}

+++ content/browser/first_party_sets/addition_overlaps_union_find.cc
@@ -15,19 +15,19 @@ namespace content {
 AdditionOverlapsUnionFind::AdditionOverlapsUnionFind(int num_sets) {
   CHECK_GE(num_sets, 0);
   representatives_.resize(num_sets);
-  std::iota(representatives_.begin(), representatives_.end(), 0);
+  std::iota(representatives_.begin(), representatives_.end(), 0ul);
 }
 
 AdditionOverlapsUnionFind::~AdditionOverlapsUnionFind() = default;
 
-void AdditionOverlapsUnionFind::Union(int set_x, int set_y) {
-  CHECK_GE(set_x, 0);
-  CHECK_LT(set_x, static_cast<int>(representatives_.size()));
-  CHECK_GE(set_y, 0);
-  CHECK_LT(set_y, static_cast<int>(representatives_.size()));
+void AdditionOverlapsUnionFind::Union(size_t set_x, size_t set_y) {
+  CHECK_GE(set_x, 0ul);
+  CHECK_LT(set_x, representatives_.size());
+  CHECK_GE(set_y, 0ul);
+  CHECK_LT(set_y, representatives_.size());
 
-  int root_x = Find(set_x);
-  int root_y = Find(set_y);
+  size_t root_x = Find(set_x);
+  size_t root_y = Find(set_y);
 
   if (root_x == root_y)
     return;
@@ -48,22 +48,22 @@ AdditionOverlapsUnionFind::SetsMap AdditionOverlapsUnionFind::SetsMapping() {
   // After the intermediate vector is populated, and we can use
   // base::MakeFlatMap to construct the mapping all at once.
   // This improvement makes this method less straightforward however.
-  for (int i = 0; i < static_cast<int>(representatives_.size()); i++) {
-    int cur_rep = Find(i);
+  for (size_t i = 0; i < representatives_.size(); i++) {
+    size_t cur_rep = Find(i);
     if (!sets.contains(cur_rep)) {
-      sets.emplace(cur_rep, base::flat_set<int>());
+      sets.emplace(cur_rep, base::flat_set<size_t>());
     }
     if (i != cur_rep) {
-      base::flat_set<int>& overlaps = sets.at(cur_rep);
+      base::flat_set<size_t>& overlaps = sets.at(cur_rep);
       overlaps.insert(i);
     }
   }
   return sets;
 }
 
-int AdditionOverlapsUnionFind::Find(int set) {
-  CHECK_GE(set, 0);
-  CHECK_LT(set, static_cast<int>(representatives_.size()));
+size_t AdditionOverlapsUnionFind::Find(size_t set) {
+  CHECK_GE(set, 0ul);
+  CHECK_LT(set, representatives_.size());
   if (representatives_[set] != set)
     representatives_[set] = Find(representatives_[set]);
   return representatives_[set];
